pyproject.toml
[project]
name = "warlock"

version = "0.1.0"

description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright",
    "httpx",
    "beautifulsoup4",
    "python-dotenv",
    "loguru>=0.7.3",
    "requests>=2.32.4",
    "discord-py>=2.4.0",
    "pyyaml>=6.0.3",
]

[project.scripts]
warlock = "fazuh.warlock.__main__:main_sync"

[project.urls]
Source = "https://github.com/FAZuH/warlock"
Issues = "https://github.com/FAZuH/warlock/issues"
Documentation = "https://github.com/FAZuH/warlock#readme"

[tool.setuptools.packages.find]
where = ["src"]
include = [ 
    "fazuh.warlock*",
]

# Common
[dependency-groups]
dev = [
    "black>=24.10.0",
    "pytest>=8.3.3",
    "isort>=5.13.2",
    "ruff>=0.7.4",
]
sandbox = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.9.2",
    "neovim>=0.3.1",
]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = [
    "tests",
]

[tool.isort]
profile = "google"
line_length = 100

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 100
indent-width = 4
target-version = "py313"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = ["E731"]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"
